Updates for magit, Emacs; New wip Emacs Lisp ports
wip/dash-el-git
- New pkg
- Synopsis: Functional operations on sets and sequences, in Emacs Lisp
- Documentation avaialble in Texinfo format

wip/emacs-git:
- Add a 'modules' option (enabled by default) thus providing support for
  libraries such as libegit2-git (used by magit, theoretically optional)
- Join PKG_SUPPORTED_OPTIONS declrations in options.mk
- Update PLIST for upstream additions (image-converter.el)

wip/git-modes-git
- Begin using a GIT_BASE parameter in the path to Emacs modules.mk -
  possibly affecting interoperability with Emacs buildlink support, also
  Makefile declrations. This is princially for a purpose of building any
  Emacs Lisp packages with either wip/emacs*git  or editors/emacs* from
  pkgsrc main
- Update PLIST for usptream changes

wip/libegit2-git
- New pkg
- Synopsis: Emacs bindings for libgit2
- Conflicts with devel/libgit2
- Published by the original develoepr of devel/libgit2
- Used by magit (theoretically, an optional component)
- Requires an Emacs built under such as the 'modules' option, as added
  to wip/emacs-git, in this changeset (and subsq, wip/emacs-nox11-git)
- Texinfo documentation available

wip/magit-git
- Update for synch with upstream changes
- Use dash-el-git, libegit2-git ports
- Add a local bundling for magit's with-editor and transient Emacs Lisp
  dependencies. Build Emacs Lisp bytecode and Texinfo documentation for
  these internal dependencies - installing as provided with this pkg -
  as well as for magit itself.

As a matter of something of an ad hoc convention, these Emacs Lisp ports
have begun using an EMACS_BASE syntax and semantics in their Makefile
and buildlink3.mk declarations. This is principally for a purpose of
making a portable  manner of reference to an Emacs modules.mk, in a
manner such that may permit for using either the 'wip' or main 'editors'
builds for Emacs or XEmacs.  As these ports are developed under wip,
they will typically use wip/emacs-git/modules.mk as a default.
 (TBD: Documentation, Discussion, Q/A)

New files in these changesets may be tagged as '$Thinkum$' in their
revision tags. This is, simply, to make no assumption for adoption at
NetBSD upstream.

The revison tags in some files modified in these changesets have been
abbreviated to '$NetBSD$' with revision tag metadata otherwise removed,
in the changeset. This is for a purpose of indicating that the file
differs from its original upstream revision.

TBD: RCS tag management for SCCM w/ Git